project_manager/requirements.txt
Django==5.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
drf_yasg==1.21.7
Requests==2.32.3
